chore: code organisation for prover service
