+++ crates/evm-prover/src/prover/message.rs
@@ -1,0 +1,5 @@
+#![allow(dead_code)]
+use sp1_sdk::include_elf;
+
+/// The ELF (executable and linkable format) file for the Succinct RISC-V zkVM.
+pub const EVM_HYPERLANE_ELF: &[u8] = include_elf!("evm-hyperlane-program");
