+++ crates/evm-prover/src/prover/mod.rs
@@ -5,8 +5,12 @@ use async_trait::async_trait;
 use sp1_sdk::{EnvProver, SP1ProofMode, SP1ProofWithPublicValues, SP1Stdin};
 
 #[allow(clippy::module_inception)]
-pub mod prover;
+pub mod block;
+pub mod message;
+pub mod range;
 pub mod service;
+
+pub use range::BlockRangeExecProver;
 
 /// ProverConfig defines metadata about the program binary (ELF), proof mode and any static keys.
 pub struct ProverConfig {
